refactor(capabilities): make tcp a submodule that owns its kinds
Closes #2160
